Rewrote PropertyGridViewMAUI tests
