--- src/System.Windows.Forms/tests/IntegrationTests/MauiTests/MauiPropertyGridViewTests/MauiPropertyGridViewTests.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <Import Project="..\References.targets"/>
-  
-  <PropertyGroup>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-  </PropertyGroup>
-
-</Project>
